Fixing loading problems
Declaring openseadragon as a dependency of openseadragon-filtering in the dependencies file of the uv-openseadron-extension. Moving rangeslider from a dependency of uv-openseadragon-extension to bundle.min.js because of its dependency of jquery. Since jquery is loaded as a CDN I couldnt get requirejs to not load jquery again and when it did it disabled the pubsub package.